+++ sdk/resourcemanager/elastic/armelastic/response_types.go
@@ -156,3 +156,8 @@ type VMHostClientListResponse struct {
 type VMIngestionClientDetailsResponse struct {
 	VMIngestionDetailsResponse
 }
+
+// VersionsClientListResponse contains the response from method VersionsClient.List.
+type VersionsClientListResponse struct {
+	VersionsListResponse
+}